More casting issues with older versions of GLib.
Try to fix

packet-btl2cap.c: In function 'btl2cap_cid_value':
packet-btl2cap.c:425: warning: cast to pointer from integer of different size
packet-btl2cap.c: In function 'btl2cap_psm_value':
packet-btl2cap.c:448: warning: cast to pointer from integer of different size

packet-btrfcomm.c: In function 'btrfcomm_directed_channel_value':
packet-btrfcomm.c:287: warning: cast to pointer from integer of different size

packet-btobex.c: In function 'btobex_profile_value':
packet-btobex.c:1134: warning: cast to pointer from integer of different size

on the 32-bit OS X builder.

Reviewed-on: https://code.wireshark.org/review/9915

+++ epan/dissectors/packet-btl2cap.c
@@ -422,7 +422,7 @@ static gpointer btl2cap_cid_value(packet_info *pinfo)
     value_data = (guint16 *) p_get_proto_data(pinfo->pool, pinfo, proto_btl2cap, PROTO_DATA_BTL2CAP_CID);
 
     if (value_data)
-        return GUINT_TO_POINTER(*value_data);
+        return GUINT_TO_POINTER((gulong)*value_data);
 
     return NULL;
 }
@@ -445,7 +445,7 @@ static gpointer btl2cap_psm_value(packet_info *pinfo)
     value_data = (guint16 *) p_get_proto_data(pinfo->pool, pinfo, proto_btl2cap, PROTO_DATA_BTL2CAP_PSM);
 
     if (value_data)
-        return GUINT_TO_POINTER(*value_data);
+        return GUINT_TO_POINTER((gulong)*value_data);
 
     return NULL;
 }
